refactor: extract operator arg helpers
